// src/generated/prisma/enums.ts
export const UserRole = {
  USER: 'USER',
  SUPER_ADMIN: 'SUPER_ADMIN',
  CONTENT_MANAGER: 'CONTENT_MANAGER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]
// ...
export const CampaignStatus = {
  DRAFT: 'DRAFT',
  SCHEDULED: 'SCHEDULED',
  SENT: 'SENT',
  FAILED: 'FAILED'
} as const

export type CampaignStatus = (typeof CampaignStatus)[keyof typeof CampaignStatus]